moved linkStorage declaration from browser transport to web sdk
